We have an existing utility stucture - use it.
